transfer code

+++ src/main/java/com/amlakie/usermanagment/dto/attendance/CarAttendanceResponseDTO.java
@@ -20,5 +20,5 @@ public class CarAttendanceResponseDTO {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     private String driverName;
-    private Double kmPerLiter;
+    private float kmPerLiter;
 }
